+++ ECommerce.WebApp/client/src/_cores/_routes/router.tsx
@@ -11,9 +11,10 @@ import {
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import DefaultLayout from "src/_shares/_layouts/default-layout/default-layout.component";
 import AdminLayout from "src/_shares/_layouts/admin-layout/admin-layout.component";
-import { 
+import {
   Login,
   Dashboard,
+  ProductList
 } from "src/_pages/admin/components";
 const Router = () => {
   const browserRoutes = createBrowserRouter([
@@ -36,7 +37,7 @@ const Router = () => {
       element: <AdminLayout />,
       children: [
         { path: "/v2/admin", element: <Dashboard /> },
-        { path: "/v2/admin/manage-product", element: <div>product</div> }
+        { path: "/v2/admin/manage-product", element: <ProductList /> }
       ],
     }
   ]);
