+++ ECommerce.WebApp/client/src/_pages/admin/components/home/dashboard.tsx
@@ -1,0 +1,5 @@
+export default function Dashboard() {
+    return (
+        <div>Home</div>
+    );
+}
