+++ ECommerce.WebApp/client/src/_pages/admin/components/index.ts
@@ -1,3 +1,3 @@
 export { default as Login } from './user/login';
-export { default as Dashboard } from './main/dashboard';
+export { default as Dashboard } from './home/dashboard';
 export { default as ProductList } from './inventory/product-list';
