+++ ECommerce.WebApp/client/src/_shares/_components/index.ts
@@ -1,3 +1,4 @@
 export { default as WebDirectional } from './web-directional/web-directional.component';
 export { default as MuiIcon } from './mui-icon/mui-icon.component';
 export { default as HeaderNavMobile } from './header-nav/nav-mobile.component';
+export { default as DataTable } from './data-table/data-table';
